Orc
Gave orc health, the ability to be damaged, and to move towards the player. Currently working on the player being able to send a fireball at it, the direction the player is facing can be detected. Also, to my knowledge, made sure the projectile of the turret is destroyed after four seconds.